micro_scalp_engine/ingest/bybit_streamer/requirements.txt
google-cloud-pubsub
google-cloud-monitoring
pybit
aiohttp 